feat(costs): account_type & provider breakdown/drill in cost explorer (DNO-358)
Registers Account Type (team/personal) and Provider (Claude/Codex/Cursor) in
the cost taxonomy (PIVOTS + DIM_ATTRIBUTE_KEY), so /costs can break down and
drill into personal vs enterprise spend and per-provider spend. They appear in
the breakdown dropdown only when the org has classified data (gated on the
gram.account_type / gram.provider attribute keys), and drilling a value ANDs it
into the query filter like any other dimension. displayName renders providers
via the shared providerLabel and capitalizes account types.

Extracts the pure DisplayAccount/providerLabel helpers out of account-display.tsx
into account-display-utils.ts so the component file satisfies the react-refresh
only-export-components rule and the taxonomy module can import the label helper.
