+++ .changeset/costs-account-type-provider.md
@@ -1,0 +1,5 @@
+---
+"dashboard": patch
+---
+
+Add Account Type (team/personal) and Provider (Claude/Codex/Cursor) as breakdown axes and drill-down filters in the cost explorer (/costs), so admins can pivot personal vs enterprise spend side by side and drill into a single account type or provider. Both appear in the breakdown dropdown only when the org has classified data.
