+++ client/dashboard/src/pages/costs/taxonomy.ts
@@ -1,4 +1,5 @@
 import { Dimension } from "@gram/client/models/components";
+import { providerLabel } from "@/components/observe/account-display-utils";
 
 // Shared cost-taxonomy config + helpers, used by both the CostsExplorer
 // controller and the EntityProfile view. Kept in a non-component module so the
@@ -46,6 +47,8 @@ export const PIVOTS: DimMeta[] = [
   { dim: Dimension.EmployeeType, label: "Employment Type" },
   { dim: Dimension.CostCenterName, label: "Cost Center" },
   { dim: Dimension.Model, label: "Model" },
+  { dim: Dimension.AccountType, label: "Account Type" },
+  { dim: Dimension.Provider, label: "Provider" },
   { dim: Dimension.Role, label: "Role" },
 ];
 
@@ -148,6 +151,8 @@ const DIM_ATTRIBUTE_KEY: Partial<Record<Dimension, string>> = {
   [Dimension.Role]: "user.roles",
   [Dimension.Model]: "gen_ai.response.model",
   [Dimension.HookSource]: "gram.hook.source",
+  [Dimension.AccountType]: "gram.account_type",
+  [Dimension.Provider]: "gram.provider",
 };
 
 // Build the set of dimensions that actually have data from the attribute keys.
@@ -202,6 +207,10 @@ export function defaultGroupBy(
 // otherwise. Shared by the profile header and the breadcrumb substitutions.
 export function displayName(dim: Dimension, value: string): string {
   if (value === "") return "(unset)";
+  if (dim === Dimension.Provider) return providerLabel(value);
+  if (dim === Dimension.AccountType) {
+    return value.charAt(0).toUpperCase() + value.slice(1);
+  }
   if (dim === Dimension.Email && value.includes("@")) {
     const local = value.split("@")[0] ?? value;
     return local
